Add a simplified version of the sdk finder.
This version is hopefully much more clear than the previous one.

The list of viable SDKs begins as either (1) all SDKs, if --sdks
contains "all" or "present", or (2) any explicitly named SDKs in --sdks.

This list is then filtered. An SDK is skipped if (1) it is not supported
by the project (eg cs2/dota for SM) or (2) it has no valid build targets
on the current platform.

Next, the SDK path is resolved. If not found:
 - If --sdks contains 'all', or the SDK was explicitly listed, then a
   fatal error is raised.
 - Otherwise, the SDK is considered optional and only a warning is
   displayed.

If an explicitly requested SDK was not found in the manifest directory,
an exception is raised.  This can happen if the SDK is mispelled.

If there are no valid build targets for any SDKs, an exception is
raised.